requirements.txt
absl-py==0.9.0
certifi==2020.4.5.2
chardet==3.0.4
future==0.18.2
idna==2.9
numpy==1.19.0
Pillow==7.1.2
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.0
six==1.15.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.46.1
urllib3==1.25.9
wget==3.2
yacs==0.1.7
